requirements.txt
torch>=1.13
torchvision
transformers>=4.20.0
Pillow
numpy
scikit-learn
streamlit
tqdm
requests
diffusers==0.23.1  
efficientnet_pytorch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git


#torch>=1.9.0
#torchvision>=0.10.0
#transformers>=4.20.0
#streamlit>=1.22.0
#Pillow>=9.0.0
#numpy>=1.21.0
#gradio>=3.0.0
#plotly>=5.5.0
